requirements.txt
streamlit==1.23.1
openai==1.12.0
pytube==15.0.0
youtube-transcript-api==0.6.2
gTTS==2.5.1
requests==2.28.2
python-dotenv==0.21.0
httpx>=0.24.1
urllib3==1.26.15 